+++ QMA.DataAccess.JsonFile/QuestionSetRepository.cs
@@ -34,7 +34,7 @@ namespace QMA.DataAccess.JsonFile
         }
 
         /// <inheritdoc/>
-        public void Add(QuestionSet value)
+        public async Task AddAsync(QuestionSet value)
         {
             if (value == null)
             {
@@ -42,7 +42,7 @@ namespace QMA.DataAccess.JsonFile
             }
 
             var coll = DataStoreSingleton.Instance.DataStore.GetCollection<QuestionSet>();
-            var success = coll.InsertOne(value);
+            var success = await coll.InsertOneAsync(value);
             if (success == false)
             {
                 throw new OperationFailedException("Add failed");
@@ -50,7 +50,7 @@ namespace QMA.DataAccess.JsonFile
         }
 
         /// <inheritdoc/>
-        public void Update(QuestionSet value)
+        public async Task UpdateAsync(QuestionSet value)
         {
             if (value == null)
             {
@@ -58,7 +58,7 @@ namespace QMA.DataAccess.JsonFile
             }
 
             var coll = DataStoreSingleton.Instance.DataStore.GetCollection<QuestionSet>();
-            var success = coll.ReplaceOne(value.PrimaryKey, value);
+            var success = await coll.ReplaceOneAsync(value.PrimaryKey, value);
             if (success == false)
             {
                 throw new OperationFailedException("Update failed");