+++ QMA.DataAccess/IAssignedQuestionRepository.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using QMA.Model.Season;
 
 namespace QMA.DataAccess
@@ -37,18 +38,21 @@ namespace QMA.DataAccess
         /// Add a new value, cannot update an existing value
         /// </summary>
         /// <param name="value">The value to add</param>
-        void Add(AssignedQuestion value);
+        /// <returns>Task for asynchronous operation</returns>
+        Task AddAsync(AssignedQuestion value);
 
         /// <summary>
         /// Delete value by primary key
         /// </summary>
         /// <param name="key">Primary key</param>
-        void Delete(string key);
+        /// <returns>Task for asynchronous operation</returns>
+        Task DeleteAsync(string key);
 
         /// <summary>
         /// Delete multiple values by Team Member ID
         /// </summary>
         /// <param name="id">Team Member ID</param>
-        void DeleteAllByTeamMemberId(string id);
+        /// <returns>Task for asynchronous operation</returns>
+        Task DeleteAllByTeamMemberIdAsync(string id);
     }
 }