+++ QMA.Helpers/AsyncHelpers.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace QMA.Helpers
+{
+    /// <summary>
+    /// Helpers for running asynchronous methods synchronously
+    /// </summary>
+    public static class AsyncHelper
+    {
+        private static readonly TaskFactory _myTaskFactory = new
+          TaskFactory(CancellationToken.None,
+                      TaskCreationOptions.None,
+                      TaskContinuationOptions.None,
+                      TaskScheduler.Default);
+
+        public static TResult RunSync<TResult>(Func<Task<TResult>> func)
+        {
+            return _myTaskFactory
+              .StartNew(func)
+              .Unwrap()
+              .GetAwaiter()
+              .GetResult();
+        }
+
+        public static void RunSync(Func<Task> func)
+        {
+            _myTaskFactory
+              .StartNew(func)
+              .Unwrap()
+              .GetAwaiter()
+              .GetResult();
+        }
+    }
+}