+++ QMA.ViewModel/EditQuestions.cs
@@ -14,6 +14,7 @@ using System.Threading.Tasks;
 using System.Windows.Input;
 using System.IO;
 using QMA.ViewModel.Services;
+using QMA.Helpers;
 
 namespace QMA.ViewModel
 {
@@ -43,9 +44,9 @@ namespace QMA.ViewModel
                     var newItem = new ObservableQuestion(
                         true,
                         item,
-                        new RelayCommand(DeleteCommand),
-                        new RelayCommand(RestoreCommand),
-                        new RelayCommand(SaveCommand)
+                        new AsyncRelayCommand(DeleteAsyncCommand),
+                        new AsyncRelayCommand(RestoreAsyncCommand),
+                        new AsyncRelayCommand(SaveAsyncCommand)
                     );
                     Items.Add(newItem);
                 }
@@ -62,9 +63,9 @@ namespace QMA.ViewModel
                         QuestionSetId = _questionSetId,
                         Text = $"Question {Items.Count + 1}"
                     },
-                    new RelayCommand(DeleteCommand),
-                    new RelayCommand(RestoreCommand),
-                    new RelayCommand(SaveCommand)
+                    new AsyncRelayCommand(DeleteAsyncCommand),
+                    new AsyncRelayCommand(RestoreAsyncCommand),
+                    new AsyncRelayCommand(SaveAsyncCommand)
                 );
                 Items.Add(newItem);
                 Selected = newItem;
@@ -88,7 +89,7 @@ namespace QMA.ViewModel
                 }
                 else
                 {
-                    SaveCommand();
+                    await SaveAsyncCommand();
                     Add.NotifyCanExecuteChanged();
                 }
             });
@@ -139,7 +140,7 @@ namespace QMA.ViewModel
 
         #endregion
 
-        private void DeleteCommand()
+        private async Task DeleteAsyncCommand()
         {
             if (Selected.Deleted != null)
             {
@@ -149,7 +150,7 @@ namespace QMA.ViewModel
             if (Selected.Persisted == true)
             {
                 Selected.Deleted = DateTimeOffset.UtcNow;
-                _repository.Update(Selected.GetModel());
+                await _repository.UpdateAsync(Selected.GetModel());
             }
             else
             {
@@ -158,7 +159,7 @@ namespace QMA.ViewModel
             Add.NotifyCanExecuteChanged();
         }
 
-        private void RestoreCommand()
+        private async Task RestoreAsyncCommand()
         {
             if (Selected.Deleted == null)
             {
@@ -166,20 +167,20 @@ namespace QMA.ViewModel
             }
 
             Selected.Deleted = null;
-            _repository.Update(Selected.GetModel());
+            await _repository.UpdateAsync(Selected.GetModel());
         }
 
-        private void SaveCommand()
+        private async Task SaveAsyncCommand()
         {
             if(Selected != null)
             {
                 if (Selected.Persisted)
                 {
-                    _repository.Update(Selected.GetModel());
+                    await _repository.UpdateAsync(Selected.GetModel());
                 }
                 else
                 {
-                    _repository.Add(Selected.GetModel());
+                    await _repository.AddAsync(Selected.GetModel());
                     Selected.Persisted = true;
                 }
             }