+++ QMA.ViewModel/EditQuizzers.cs
@@ -1,12 +1,14 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
 using Microsoft.Toolkit.Mvvm.Input;
 using QMA.DataAccess;
+using QMA.Helpers;
 using QMA.Model;
 using QMA.ViewModel.Observables;
 using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
+using System.Threading.Tasks;
 using System.Windows.Input;
 
 namespace QMA.ViewModel
@@ -28,9 +30,9 @@ namespace QMA.ViewModel
                     var newItem = new ObservableQuizzer(
                         true,
                         item,
-                        new RelayCommand(DeleteCommand),
-                        new RelayCommand(RestoreCommand),
-                        new RelayCommand(SaveCommand)
+                        new AsyncRelayCommand(DeleteAsyncCommand),
+                        new AsyncRelayCommand(RestoreAsyncCommand),
+                        new AsyncRelayCommand(SaveAsyncCommand)
                     );
                     Items.Add(newItem);
                 }
@@ -49,9 +51,9 @@ namespace QMA.ViewModel
                         LastName = null,
                         ParentFullName = null
                     },
-                    new RelayCommand(DeleteCommand),
-                    new RelayCommand(RestoreCommand),
-                    new RelayCommand(SaveCommand)
+                    new AsyncRelayCommand(DeleteAsyncCommand),
+                    new AsyncRelayCommand(RestoreAsyncCommand),
+                    new AsyncRelayCommand(SaveAsyncCommand)
                 );
                 Items.Add(newItem);
                 Selected = newItem;
@@ -76,7 +78,7 @@ namespace QMA.ViewModel
                 }
                 else
                 {
-                    SaveCommand();
+                    await SaveAsyncCommand();
                     Add.NotifyCanExecuteChanged();
                 }
             });
@@ -99,7 +101,7 @@ namespace QMA.ViewModel
 
         #endregion
 
-        private void DeleteCommand()
+        private async Task DeleteAsyncCommand()
         {
             if(Selected.Deleted != null)
             {
@@ -109,7 +111,7 @@ namespace QMA.ViewModel
             if (Selected.Persisted == true)
             {
                 Selected.Deleted = DateTimeOffset.UtcNow;
-                _repository.Update(Selected.GetModel());
+                await _repository.UpdateAsync(Selected.GetModel());
             }
             else
             {
@@ -118,7 +120,7 @@ namespace QMA.ViewModel
             Add.NotifyCanExecuteChanged();
         }
 
-        private void RestoreCommand()
+        private async Task RestoreAsyncCommand()
         {
             if (Selected.Deleted == null)
             {
@@ -126,20 +128,20 @@ namespace QMA.ViewModel
             }
 
             Selected.Deleted = null;
-            _repository.Update(Selected.GetModel());
+            await _repository.UpdateAsync(Selected.GetModel());
         }
 
-        private void SaveCommand()
+        private async Task SaveAsyncCommand()
         {
             if(Selected != null)
             {
                 if (Selected.Persisted)
                 {
-                    _repository.Update(Selected.GetModel());
+                    await _repository.UpdateAsync(Selected.GetModel());
                 }
                 else
                 {
-                    _repository.Add(Selected.GetModel());
+                    await _repository.AddAsync(Selected.GetModel());
                     Selected.Persisted = true;
                 }
             }