+++ QMA.ViewModel/EditSeasons.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Toolkit.Mvvm.ComponentModel;
 using Microsoft.Toolkit.Mvvm.Input;
 using QMA.DataAccess;
+using QMA.Helpers;
 using QMA.Model;
 using QMA.ViewModel.Observables;
 using System;
@@ -42,9 +43,9 @@ namespace QMA.ViewModel
                     var newItem = new ObservableSeason(
                         true,
                         item,
-                        new RelayCommand(DeleteCommand),
-                        new RelayCommand(RestoreCommand),
-                        new RelayCommand(SaveCommand)
+                        new AsyncRelayCommand(DeleteAsyncCommand),
+                        new AsyncRelayCommand(RestoreAsyncCommand),
+                        new AsyncRelayCommand(SaveAsyncCommand)
                     );
                     Items.Add(newItem);
                 }
@@ -60,9 +61,9 @@ namespace QMA.ViewModel
                         PrimaryKey = Guid.NewGuid().ToString(),
                         Name = $"Season {Items.Count + 1}",
                     },
-                    new RelayCommand(DeleteCommand),
-                    new RelayCommand(RestoreCommand),
-                    new RelayCommand(SaveCommand)
+                    new AsyncRelayCommand(DeleteAsyncCommand),
+                    new AsyncRelayCommand(RestoreAsyncCommand),
+                    new AsyncRelayCommand(SaveAsyncCommand)
                 );
                 Items.Add(newItem);
                 Selected = newItem;
@@ -86,7 +87,7 @@ namespace QMA.ViewModel
                 }
                 else
                 {
-                    SaveCommand();
+                    await SaveAsyncCommand();
                     Add.NotifyCanExecuteChanged();
                 }
             });
@@ -111,7 +112,7 @@ namespace QMA.ViewModel
 
         #endregion
 
-        private void DeleteCommand()
+        private async Task DeleteAsyncCommand()
         {
             if (Selected.Deleted != null)
             {
@@ -121,7 +122,7 @@ namespace QMA.ViewModel
             if (Selected.Persisted == true)
             {
                 Selected.Deleted = DateTimeOffset.UtcNow;
-                _repository.Update(Selected.GetModel());
+                await _repository.UpdateAsync(Selected.GetModel());
             }
             else
             {
@@ -130,7 +131,7 @@ namespace QMA.ViewModel
             Add.NotifyCanExecuteChanged();
         }
 
-        private void RestoreCommand()
+        private async Task RestoreAsyncCommand()
         {
             if (Selected.Deleted == null)
             {
@@ -138,20 +139,20 @@ namespace QMA.ViewModel
             }
 
             Selected.Deleted = null;
-            _repository.Update(Selected.GetModel());
+            await _repository.UpdateAsync(Selected.GetModel());
         }
 
-        private void SaveCommand()
+        private async Task SaveAsyncCommand()
         {
             if(Selected != null)
             {
                 if (Selected.Persisted)
                 {
-                    _repository.Update(Selected.GetModel());
+                    await _repository.UpdateAsync(Selected.GetModel());
                 }
                 else
                 {
-                    _repository.Add(Selected.GetModel());
+                    await _repository.AddAsync(Selected.GetModel());
                     Selected.Persisted = true;
                 }
             }