+++ QuizMastersApprenticeApp/App.xaml.cs
@@ -31,6 +31,8 @@ namespace QuizMastersApprenticeApp
             PresentationTraceSources.DataBindingSource.Listeners.Add(new DebugTraceListener());
 #endif
 
+            SetupUnhandledExceptionHandling();
+
             SetupAppDataFolder();
             SetupCommonDocFolder();
             SetupUserDocFolder();
@@ -60,6 +62,47 @@ namespace QuizMastersApprenticeApp
             UserDocFolder = Path.Combine(UserDocFolder, AppName);
 
             Directory.CreateDirectory(UserDocFolder);
+        }
+
+        private void SetupUnhandledExceptionHandling()
+        {
+            // Catch exceptions from all threads in the AppDomain
+            AppDomain.CurrentDomain.UnhandledException += (sender, args) =>
+                ShowUnhandledException(args.ExceptionObject as Exception, "AppDomain.CurrentDomain.UnhandledException", false);
+
+            // Catch exceptions from each AppDomain that uses a task scheduler for async operations
+            TaskScheduler.UnobservedTaskException += (sender, args) =>
+                ShowUnhandledException(args.Exception, "TaskScheduler.UnobservedTaskException", false);
+
+            // Catch exceptions from a single specific UI dispatcher thread
+            Dispatcher.UnhandledException += (sender, args) =>
+            {
+                // If we are debugging, let Visual Studio handle the exception and take us to the code that threw it.
+                if (!Debugger.IsAttached)
+                {
+                    args.Handled = true;
+                    ShowUnhandledException(args.Exception, "Dispatcher.UnhandledException", true);
+                }
+            };
+        }
+
+        private void ShowUnhandledException(Exception e, string unhandledExceptionType, bool promptUserForShutdown)
+        {
+            var messageBoxTitle = $"An unexpected error occurred: {unhandledExceptionType}";
+            var messageBoxMessage = $"Exception information:\n\n{e}";
+            var messageBoxButtons = MessageBoxButton.OK;
+
+            if (promptUserForShutdown)
+            {
+                messageBoxMessage += "\n\nIt is recommended that the application be closed. Do you want to close the application?";
+                messageBoxButtons = MessageBoxButton.YesNo;
+            }
+
+            // Let the user decide if the app should die or not (if applicable).
+            if (MessageBox.Show(messageBoxMessage, messageBoxTitle, messageBoxButtons, MessageBoxImage.Error) == MessageBoxResult.Yes)
+            {
+                Current.Shutdown();
+            }
         }
     }